requirements.txt
requests
requests-toolbelt
beautifulsoup4
GoogleAppEngineCloudStorageClient==1.9.22.1
